Add tests for native params
# TODO:
# - min/maxing will require field combinators in the modules lmao
