+++ tests/test_osc_native_params.py
@@ -29,11 +29,8 @@ from tests import EyeInfoMock, SimpleUDPClientMock
                 ),
             ],
             [
-                ("/avatar/parameters/LeftEyeX", 0),
-                ("/avatar/parameters/RightEyeX", 0),
-                ("/avatar/parameters/EyesY", 0),
-                ("/avatar/parameters/LeftEyeLidExpandedSqueeze", 1.0),
-                ("/avatar/parameters/RightEyeLidExpandedSqueeze", 1.0),
+                ("/tracking/eye/EyesClosedAmount", 0.0),
+                ("/tracking/eye/LeftRightVec", [0.0, 0.0, 1.0, 0.0, 0.0, 1.0]),
             ],
         ),
         (
@@ -53,20 +50,17 @@ from tests import EyeInfoMock, SimpleUDPClientMock
                 ),
             ],
             [
-                ("/avatar/parameters/LeftEyeX", 0),
-                ("/avatar/parameters/RightEyeX", 0),
-                ("/avatar/parameters/EyesY", 0),
-                ("/avatar/parameters/LeftEyeLidExpandedSqueeze", 1.0),
-                ("/avatar/parameters/RightEyeLidExpandedSqueeze", 1.0),
+                ("/tracking/eye/EyesClosedAmount", 0.0),
+                ("/tracking/eye/LeftRightVec", [0.0, 0.0, 1.0, 0.0, 0.0, 1.0]),
             ],
         ),
     ],
 )
-def test_send_command_v1_params_single_eye(main_config_v1_params, messages, expected_outcome):
+def test_send_command_native_params_single_eye(main_config_native_params, messages, expected_outcome):
     with mock.patch("EyeTrackApp.osc.osc.udp_client.SimpleUDPClient", SimpleUDPClientMock):
         msg_queue = Queue()
         client = OSCManager(
-            config=main_config_v1_params,
+            config=main_config_native_params,
             osc_message_in_queue=msg_queue,
         )
 
@@ -114,13 +108,16 @@ def test_send_command_v1_params_single_eye(main_config_v1_params, messages, expe
                 ),
             ],
             [
-                ("/avatar/parameters/RightEyeLidExpandedSqueeze", 1.0),
-                ("/avatar/parameters/RightEyeX", 0),
-                ("/avatar/parameters/RightEyeLidExpandedSqueeze", 1.0),
-                ("/avatar/parameters/LeftEyeLidExpandedSqueeze", 0.5),
-                ("/avatar/parameters/LeftEyeX", 0),
-                ("/avatar/parameters/LeftEyeLidExpandedSqueeze", 0.5),
-                ("/avatar/parameters/EyesY", 2.5),
+                ("/tracking/eye/EyesClosedAmount", 0.0),
+                ("/tracking/eye/EyesClosedAmount", 0.0),
+                # we're expecting 621 as left_y here because that's the default value
+                # before the first state update with real data, but that's ok
+                # we're gonna be like 10 messages deep before anyone starts playing
+                # and if they already are, they won't be able to notice
+                ("/tracking/eye/LeftRightVec", [0.0, 621.0, 1.0, 0.0, 0.0, 1.0]),
+                ("/tracking/eye/EyesClosedAmount", 0.5),
+                ("/tracking/eye/EyesClosedAmount", 0.5),
+                ("/tracking/eye/LeftRightVec", [0.0, 5.0, 1.0, 0.0, 0.0, 1.0]),
             ],
         ),
         # binary blink
@@ -154,34 +151,31 @@ def test_send_command_v1_params_single_eye(main_config_v1_params, messages, expe
                 ),
             ],
             [
-                ("/avatar/parameters/RightEyeLidExpandedSqueeze", 0.0),
-                ("/avatar/parameters/RightEyeLidExpandedSqueeze", 0.0),
-                ("/avatar/parameters/RightEyeLidExpandedSqueeze", 0.0),
-                ("/avatar/parameters/RightEyeLidExpandedSqueeze", 0.0),
-                ("/avatar/parameters/RightEyeLidExpandedSqueeze", 0.0),
-                ("/avatar/parameters/RightEyeLidExpandedSqueeze", 0.0),
-                ("/avatar/parameters/RightEyeX", 0),
-                ("/avatar/parameters/RightEyeLidExpandedSqueeze", 0.0),
-                ("/avatar/parameters/LeftEyeLidExpandedSqueeze", 0.0),
-                ("/avatar/parameters/LeftEyeLidExpandedSqueeze", 0.0),
-                ("/avatar/parameters/LeftEyeLidExpandedSqueeze", 0.0),
-                ("/avatar/parameters/LeftEyeLidExpandedSqueeze", 0.0),
-                ("/avatar/parameters/LeftEyeLidExpandedSqueeze", 0.0),
-                ("/avatar/parameters/LeftEyeLidExpandedSqueeze", 0.0),
-                ("/avatar/parameters/LeftEyeX", 0),
-                ("/avatar/parameters/LeftEyeLidExpandedSqueeze", 0.0),
-                ("/avatar/parameters/EyesY", 2.5),
+                ("/tracking/eye/EyesClosedAmount", 1.0),
+                ("/tracking/eye/EyesClosedAmount", 1.0),
+                ("/tracking/eye/EyesClosedAmount", 1.0),
+                ("/tracking/eye/EyesClosedAmount", 1.0),
+                ("/tracking/eye/EyesClosedAmount", 1.0),
+                ("/tracking/eye/EyesClosedAmount", 1.0),
+                ("/tracking/eye/LeftRightVec", [0.0, 621.0, 1.0, 0.0, 0.0, 1.0]),
+                ("/tracking/eye/EyesClosedAmount", 1.0),
+                ("/tracking/eye/EyesClosedAmount", 1.0),
+                ("/tracking/eye/EyesClosedAmount", 1.0),
+                ("/tracking/eye/EyesClosedAmount", 1.0),
+                ("/tracking/eye/EyesClosedAmount", 1.0),
+                ("/tracking/eye/EyesClosedAmount", 1.0),
+                ("/tracking/eye/LeftRightVec", [0.0, 5.0, 1.0, 0.0, 0.0, 1.0]),
             ],
         ),
     ],
 )
-def test_send_command_v1_params_dual_eye(main_config_v1_params, eye_data, expected_outcome):
-    main_config_v1_params.eye_display_id = 2
+def test_send_command_native_params_dual_eye(main_config_native_params, eye_data, expected_outcome):
+    main_config_native_params.eye_display_id = 2
 
     with mock.patch("EyeTrackApp.osc.osc.udp_client.SimpleUDPClient", SimpleUDPClientMock):
         msg_queue = Queue()
         client = OSCManager(
-            config=main_config_v1_params,
+            config=main_config_native_params,
             osc_message_in_queue=msg_queue,
         )
 
@@ -229,37 +223,38 @@ def test_send_command_v1_params_dual_eye(main_config_v1_params, eye_data, expect
                 ),
             ],
             [
-                ("/avatar/parameters/RightEyeLidExpandedSqueeze", 0.0),
-                ("/avatar/parameters/RightEyeLidExpandedSqueeze", 0.0),
-                ("/avatar/parameters/RightEyeLidExpandedSqueeze", 0.0),
-                ("/avatar/parameters/RightEyeLidExpandedSqueeze", 0.0),
-                ("/avatar/parameters/RightEyeLidExpandedSqueeze", 0.0),
-                ("/avatar/parameters/RightEyeLidExpandedSqueeze", 0.0),
-                ("/avatar/parameters/RightEyeX", 0),
-                ("/avatar/parameters/RightEyeLidExpandedSqueeze", 0.0),
-                ("/avatar/parameters/LeftEyeLidExpandedSqueeze", 0.0),
-                ("/avatar/parameters/LeftEyeLidExpandedSqueeze", 0.0),
-                ("/avatar/parameters/LeftEyeLidExpandedSqueeze", 0.0),
-                ("/avatar/parameters/LeftEyeLidExpandedSqueeze", 0.0),
-                ("/avatar/parameters/LeftEyeLidExpandedSqueeze", 0.0),
-                ("/avatar/parameters/LeftEyeLidExpandedSqueeze", 0.0),
-                ("/avatar/parameters/RightEyeLidExpandedSqueeze", 0.0),
-                ("/avatar/parameters/LeftEyeLidExpandedSqueeze", 0.0),
-                ("/avatar/parameters/LeftEyeX", 0),
-                ("/avatar/parameters/LeftEyeLidExpandedSqueeze", 0.0),
-                ("/avatar/parameters/EyesY", 2.5),
+                ("/tracking/eye/EyesClosedAmount", 1.0),
+                ("/tracking/eye/EyesClosedAmount", 1.0),
+                ("/tracking/eye/EyesClosedAmount", 1.0),
+                ("/tracking/eye/EyesClosedAmount", 1.0),
+                ("/tracking/eye/EyesClosedAmount", 1.0),
+                ("/tracking/eye/EyesClosedAmount", 1.0),
+                ("/tracking/eye/LeftRightVec", [0.0, 621.0, 1.0, 0.0, 0.0, 1.0]),
+                ("/tracking/eye/EyesClosedAmount", 1.0),
+                ("/tracking/eye/EyesClosedAmount", 1.0),
+                ("/tracking/eye/EyesClosedAmount", 1.0),
+                ("/tracking/eye/EyesClosedAmount", 1.0),
+                ("/tracking/eye/EyesClosedAmount", 1.0),
+                ("/tracking/eye/EyesClosedAmount", 1.0),
+                ("/tracking/eye/EyesClosedAmount", 1.0),
+                ("/tracking/eye/EyesClosedAmount", 1.0),
+                ("/tracking/eye/EyesClosedAmount", 1.0),
+                ("/tracking/eye/EyesClosedAmount", 1.0),
+                ("/tracking/eye/EyesClosedAmount", 1.0),
+                ("/tracking/eye/EyesClosedAmount", 1.0),
+                ("/tracking/eye/LeftRightVec", [0.0, 5.0, 1.0, 0.0, 0.0, 1.0]),
             ],
         ),
     ],
 )
-def test_send_command_v1_params_eye_outer_side_falloff(main_config_v1_params, eye_data, expected_outcome):
-    main_config_v1_params.eye_display_id = 2
-    main_config_v1_params.settings.gui_outer_side_falloff = True
+def test_send_command_native_params_eye_outer_side_falloff(main_config_native_params, eye_data, expected_outcome):
+    main_config_native_params.eye_display_id = 2
+    main_config_native_params.settings.gui_outer_side_falloff = True
 
     with mock.patch("EyeTrackApp.osc.osc.udp_client.SimpleUDPClient", SimpleUDPClientMock):
         msg_queue = Queue()
         client = OSCManager(
-            config=main_config_v1_params,
+            config=main_config_native_params,
             osc_message_in_queue=msg_queue,
         )
