+++ tests/test_osc_v2_params.py
@@ -1,4 +1,3 @@
-import dataclasses
 from queue import Queue
 from time import sleep
 from unittest import mock
@@ -7,25 +6,7 @@ import pytest
 
 from osc.osc import OSCManager, OSCMessage
 from osc.OSCMessage import OSCMessageType
-
-
-@dataclasses.dataclass
-class EyeInfoMock:
-    x: int
-    y: int
-    blink: float
-    pupil_dilation: float
-    avg_velocity: float
-
-
-class SimpleUDPClientMock:
-    def __init__(self, osc_address, port):
-        self.osc_address = osc_address
-        self.port = port
-        self.messages = []
-
-    def send_message(self, address, value):
-        self.messages.append((address, value))
+from tests import EyeInfoMock, SimpleUDPClientMock
 
 
 @pytest.mark.parametrize(
